Retire the mobileEditPraxis surface and its eight files (#1181)
ADR-0065 §2, on ADR-0063's terms rather than ADR-0056/0058's: the retirement is
outright, not a dormant revert path, because all nine composer designs were
committed before a single archetype was rebuilt. The mobile skins are superseded
by a committed design, not held open pending one.

Gone: the FactionManifest field, the SURFACE_KEYS entry, all seven faction
registrations, pages/editPraxis/mobileArchetypes/ and its four test files.
EditPraxis.tsx no longer reads useFormFactor at all — each archetype calls it
internally for its own size set.

The BREADCRUMB condition is reworked, not merely trimmed. It read
`formFactor === "mobile" || waiting`, where the mobile half existed because
mobile skins painted none (#567). The desktop archetype now serves the phone and
paints its own, so that half is already covered and keeping it would draw a
SECOND breadcrumb above every mobile composer; the gate is now `waiting` alone.
composerDispatch.test.tsx counts <nav> at both widths so neither zero nor two can
come back.

That test also replaces the two the retirement took with it. It renders through
the dispatcher at BOTH widths — the claim the collapse actually makes — and
keeps the mode gates (#311's level gate, #877's locked collapse) under test. The
surfaceDispatch registry swaps its mobileEditPraxis row for an `editPraxis` one
so the composer is not left the one big surface with no row.
